requirements.txt
# Core dependencies for Ultimate Japanese Quiz Solver - Phase 2A Enhanced

# GUI and system integration
tkinter-tooltip==2.2.2
mss==9.0.1
Pillow==10.4.0
psutil==5.9.8

# OCR and image processing - Phase 2A Enhanced
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.12.0
scikit-image==0.22.0
scikit-learn==1.4.1  # Phase 2A: Enhanced ML capabilities

# AI providers
google-generativeai==0.4.1
openai==1.12.0
anthropic==0.18.1
requests==2.31.0

# Morphological analysis - Japanese NLP
fugashi[unidic-lite]==1.3.0
MeCab-python3==1.0.8
spacy==3.7.4

# Machine Learning and Transformers
transformers==4.38.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
accelerate==0.27.2
peft==0.9.0
bitsandbytes==0.42.0
datasets==2.17.1

# Vector database and embeddings - RAG
faiss-cpu==1.8.0
sentence-transformers==2.5.1
chromadb==0.4.24
langchain==0.1.13
langchain-community==0.0.29

# Text processing and similarity
nltk==3.8.1
Levenshtein==0.25.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# Database and utilities
pandas==2.2.1
jsonschema==4.21.1

# Web scraping and data collection
beautifulsoup4==4.12.3
requests-html==0.10.0

# Testing and evaluation
pytest==8.1.1
pytest-asyncio==0.23.6

# Visualization and analysis
matplotlib==3.8.3
seaborn==0.13.2
jupyter==1.0.0
notebook==7.1.2

# Configuration management
python-dotenv==1.0.1
pyyaml==6.0.1

# Optional: for advanced features
keyboard==0.13.5  # Global hotkeys

# Audio processing (for listening questions)
librosa==0.10.1
speech-recognition==3.10.1
pydub==0.25.1

# Development and debugging
tqdm==4.66.2

# Export formats
openpyxl==3.1.2
genanki==0.13.0  # Anki deck creation

# Quality assurance
flake8==7.0.0
black==24.2.0

